Implement multiple images for wikipedia
